io.sphere.sdk.queries.PagedResult and descendants as interface #1053
