+++ commercetools-sdk-base/src/main/java/io/sphere/sdk/queries/PagedQueryResultDsl.java
@@ -2,7 +2,7 @@ package io.sphere.sdk.queries;
 
 import java.util.List;
 
-public final class PagedQueryResultDsl<T> extends PagedQueryResult<T> {
+public final class PagedQueryResultDsl<T> extends PagedQueryResultImpl<T> {
     PagedQueryResultDsl(final Long offset, final Long total, final List<T> results, final Long count) {
         super(offset, total, results, count);
     }
