+++ commercetools-sdk-base/src/main/java/io/sphere/sdk/queries/PagedQueryResultImpl.java
@@ -1,0 +1,13 @@
+package io.sphere.sdk.queries;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+
+import java.util.List;
+
+class PagedQueryResultImpl<T> extends PagedResultBase<T> implements PagedQueryResult<T> {
+
+    @JsonCreator
+    PagedQueryResultImpl(final Long offset, final Long total, final List<T> results, final Long count) {
+        super(offset, total, results, count);
+    }
+}
