+++ commercetools-sdk-base/src/main/java/io/sphere/sdk/queries/PagedResult.java
@@ -1,30 +1,14 @@
 package io.sphere.sdk.queries;
 
-import io.sphere.sdk.models.Base;
-
 import java.util.List;
 import java.util.Optional;
 
-public abstract class PagedResult<T> extends Base {
-    protected final Long offset;
-    protected final Long total;
-    protected final List<T> results;
-    protected final Long count;
-
-    protected PagedResult(final Long offset, final Long total, final List<T> results, final Long count) {
-        this.offset = offset;
-        this.total = total;
-        this.results = results;
-        this.count = count;
-    }
-
-    /**
+public interface PagedResult<T> {
+     /**
       * The offset supplied by the client or the server default.
       * @return the amount of items (not pages) skipped
       */
-     public Long getOffset() {
-         return offset;
-     }
+     Long getOffset();
 
     /**
      * The actual number of results returned.
@@ -32,26 +16,20 @@ public abstract class PagedResult<T> extends Base {
      * @deprecated use {@link #getCount()} instead
      */
     @Deprecated//don't remove soon!!!
-    public Long size() {
-        return (long) results.size();
-    }
+    Long size();
 
     /**
      * The actual number of results returned.
      * @return the number of elements in this container
      */
-    public Long getCount() {
-        return count;
-    }
+    Long getCount();
 
     /**
       * The total number of results matching the request.
       * Total is greater or equal to count.
       * @return the number of elements that can be fetched matching the criteria
       */
-     public Long getTotal() {
-         return total;
-     }
+    Long getTotal();
 
     /**
       * List of results. If {@link io.sphere.sdk.queries.PagedResult#getCount()} is not equal
@@ -59,24 +37,23 @@ public abstract class PagedResult<T> extends Base {
       * elements that match the request.
       * @return {@link io.sphere.sdk.queries.PagedResult#getCount()} elements matching the request
       */
-     public List<T> getResults() {
-         return results;
-     }
+     List<T> getResults();
 
     /**
       * Tries to access the first element of the result list.
       * Use case: query by slug which should contain zero or one element in the result list.
       * @return the first value or absent
       */
-     public Optional<T> head() {
-         return results.isEmpty() ? Optional.<T>empty() : Optional.of(results.get(0));
+    default Optional<T> head() {
+        final List<T> results = getResults();
+        return results.isEmpty() ? Optional.empty() : Optional.of(results.get(0));
      }
 
     /**
       * Checks if this is the first page of a result.
       * @return true if offset is 0 otherwise false
       */
-     public boolean isFirst() {
+    default boolean isFirst() {
          if (getOffset() == null) {
              throw new UnsupportedOperationException("Can only be used if the offset is known.");
          }
@@ -87,7 +64,7 @@ public abstract class PagedResult<T> extends Base {
       * Checks if it is the last possible page.
       * @return true if doing a request with an incremented offset parameter would cause an empty result otherwise false.
       */
-     public boolean isLast() {
+     default boolean isLast() {
          if (getOffset() == null || getTotal() == null) {
              throw new UnsupportedOperationException("Can only be used if the offset & total is known.");
          }
