+++ commercetools-sdk-base/src/main/java/io/sphere/sdk/queries/PagedResultBase.java
@@ -1,0 +1,64 @@
+package io.sphere.sdk.queries;
+
+import io.sphere.sdk.models.Base;
+
+import java.util.List;
+
+public abstract class PagedResultBase<T> extends Base implements PagedResult<T> {
+    protected final Long offset;
+    protected final Long total;
+    protected final List<T> results;
+    protected final Long count;
+
+    protected PagedResultBase(final Long offset, final Long total, final List<T> results, final Long count) {
+        this.offset = offset;
+        this.total = total;
+        this.results = results;
+        this.count = count;
+    }
+
+    /**
+     * The offset supplied by the client or the server default.
+     * @return the amount of items (not pages) skipped
+     */
+    public Long getOffset() {
+        return offset;
+    }
+
+    /**
+     * The actual number of results returned.
+     * @return the number of elements in this container
+     * @deprecated use {@link #getCount()} instead
+     */
+    @Deprecated//don't remove soon!!!
+    public Long size() {
+        return (long) results.size();
+    }
+
+    /**
+     * The actual number of results returned.
+     * @return the number of elements in this container
+     */
+    public Long getCount() {
+        return count;
+    }
+
+    /**
+     * The total number of results matching the request.
+     * Total is greater or equal to count.
+     * @return the number of elements that can be fetched matching the criteria
+     */
+    public Long getTotal() {
+        return total;
+    }
+
+    /**
+     * List of results. If {@link PagedResultBase#getCount()} is not equal
+     * to {@link PagedResultBase#getTotal()} the container contains only a subset of all
+     * elements that match the request.
+     * @return {@link PagedResultBase#getCount()} elements matching the request
+     */
+    public List<T> getResults() {
+        return results;
+    }
+}
